isolando i18n da main
